improved !

+++ serial/main.c
@@ -22,7 +22,7 @@ int main()
 {
 	// srand( static_cast<unsigned int>(time(NULL)));
     // srand( (unsigned int) time(NULL) );
-	float start,stop;
+	double start,stop;
 	start = getTime();
     Network network;
     initNetwork(&network);
@@ -35,7 +35,7 @@ int main()
         testNetwork(&network);
     }
     stop = getTime();
-    float serialTime= stop-start;
+    double serialTime= stop-start;
     printf("Serial Implementation of NN\n");
     printf("\nTime Elapsed=%f",serialTime );
     return 0;
